+++ src/App/Exercises/HtmlCss/Backgrounds/index.jsx
@@ -1,0 +1,11 @@
+import './styles.css';
+
+export function Exercise() {
+  return (
+    <div>
+      <img width="200" src="./tree.jpg" alt="Zdjęcie" />
+
+      <div className="background-url"></div>
+    </div>
+  );
+}
