+++ src/App/Exercises/HtmlCss/Backgrounds/router-data.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { Exercise } from './index';
+
+export const blockRouterMetaData = {
+  path: 'backgrounds',
+  date: '12-12-2023',
+  linkLabel: 'Tła',
+  blockNo: 1,
+  element: <Exercise />,
+  tags: ['css', 'backgrounds'],
+};
