+++ src/App/Exercises/HtmlCss/Backgrounds/styles.css
@@ -1,0 +1,14 @@
+.font-test-family {
+  font-family: Chilanka;
+}
+.background-url {
+
+  background: url('http://localhost:3000/tree.jpg');
+
+  width: 500px;
+
+  height: 300px;
+
+  background-size: cover;
+
+}
