+++ src/App/Exercises/HtmlCss/Exercise-example-12-12-2023-1 copy/index.jsx
@@ -1,0 +1,5 @@
+import './styles.css';
+
+export function Exercise() {
+  return <p>Ćwiczenia z HTML & CSS - 1</p>;
+}
