+++ src/App/Exercises/HtmlCss/Exercise-example-12-12-2023-1 copy/router-data.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { Exercise } from './index';
+
+export const blockRouterMetaData = {
+  path: '12-12-2023-1',
+  date: '12-12-2023',
+  linkLabel: 'HTML & CSS - ćwiczenie 1 z dnia 12-12-2023',
+  blockNo: 1,
+  element: <Exercise />,
+  tags: ['flex', 'grid', 'video'],
+};
