+++ src/App/Exercises/HtmlCss/Exercise-example-12-12-2023-1 copy/styles.css
@@ -1,0 +1,3 @@
+.font-test-family {
+  font-family: Chilanka;
+}
