--- src/App/Exercises/HtmlCss/Exercise-example-12-12-2023-1/index.jsx
@@ -1,5 +1,0 @@
-import './styles.css';
-
-export function Exercise() {
-  return <p>Ćwiczenia z HTML & CSS - 1</p>;
-}
