--- src/App/Exercises/HtmlCss/Exercise-example-12-12-2023-1/router-data.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import { Exercise } from './index';
-
-export const blockRouterMetaData = {
-  path: '12-12-2023-1',
-  date: '12-12-2023',
-  linkLabel: 'HTML & CSS - ćwiczenie 1 z dnia 12-12-2023',
-  blockNo: 1,
-  element: <Exercise />,
-  tags: ['flex', 'grid', 'video'],
-};
