--- src/App/Exercises/HtmlCss/Exercise-example-12-12-2023-1/styles.css
@@ -1,3 +1,0 @@
-.font-test-family {
-  font-family: Chilanka;
-}
