+++ src/App/Exercises/HtmlCss/ImageFiles/router-data.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { Exercise } from './index';
+
+export const blockRouterMetaData = {
+  path: 'ImageFiles',
+  date: '12-12-2023',
+  linkLabel: 'ImageFiles',
+  blockNo: 1,
+  element: <Exercise />,
+  tags: ['flex', 'grid', 'video'],
+};
