+++ src/App/Exercises/HtmlCss/ImageFiles/styles.css
@@ -1,0 +1,3 @@
+.font-test-family {
+  font-family: Chilanka;
+}
